// src/app/shared/bootstrap-input/models/bootstrap-input-types.model.ts
export enum BootstrapInputTypes {
    Input = 'Input',
    InputFile = 'File',
    TextArea = 'TextArea',
    Select = 'Select',
    CheckBox = 'CheckBox',
    Radio = 'Radio',
}

export enum Type {
    Text = 'text',
    Email = 'email',
    Password = 'password',
    File = 'file',
    CheckBox = 'checkbox',
    Radio = 'radio',
}